# apps/api/app/routers/health.py
from fastapi import APIRouter
from pydantic import BaseModel

from app import __version__

router = APIRouter(tags=["system"])


class Health(BaseModel):
    status: str
    version: str


@router.get("/health", response_model=Health)
def health() -> Health:
    """Liveness probe."""
    return Health(status="ok", version=__version__)