# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "properties": {
          "question": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeInfo"
          }
        },
        "required": [
          "question",
          "scope"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
      "Body_upload_video_v1_videos_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_video_v1_videos_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Health": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "version"
        ],
        "title": "Health",
        "type": "object"
      },
      "JobCreate": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          }
        },
        "required": [
          "kind"
        ],
        "title": "JobCreate",
        "type": "object"
      },
      "JobKind": {
        "enum": [
          "probe",
          "indexing"
        ],
        "title": "JobKind",
        "type": "string"
      },
      "JobOut": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "progress": {
            "title": "Progress",
            "type": "number"
          },
          "startedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "videoId": {
            "title": "Videoid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "videoId",
          "kind",
          "status",
          "progress",
          "createdAt"
        ],
        "title": "JobOut",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "ScopeInfo": {
        "properties": {
          "duration": {
            "minimum": 0.0,
            "title": "Duration",
            "type": "number"
          },
          "end": {
            "minimum": 0.0,
            "title": "End",
            "type": "number"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "start": {
            "minimum": 0.0,
            "title": "Start",
            "type": "number"
          },
          "videoId": {
            "minLength": 1,
            "title": "Videoid",
            "type": "string"
          },
          "videoName": {
            "title": "Videoname",
            "type": "string"
          }
        },
        "required": [
          "videoId",
          "videoName",
          "start",
          "end",
          "duration",
          "label"
        ],
        "title": "ScopeInfo",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VideoOut": {
        "properties": {
          "codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codec"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "durationSec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationsec"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sizeBytes": {
            "title": "Sizebytes",
            "type": "integer"
          },
          "src": {
            "description": "Path (relative to the API origin) the player loads the video from.",
            "readOnly": true,
            "title": "Src",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/VideoStatus"
          },
          "thumbnailSrc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "readOnly": true,
            "title": "Thumbnailsrc"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "name",
          "filename",
          "sizeBytes",
          "status",
          "createdAt",
          "src",
          "thumbnailSrc"
        ],
        "title": "VideoOut",
        "type": "object"
      },
      "VideoStatus": {
        "enum": [
          "uploaded",
          "ready",
          "failed"
        ],
        "title": "VideoStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "LETS Video Analyzer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/analyze": {
      "post": {
        "description": "Answer a question about a video as a Server-Sent Events stream.\n\nCurrently simulated (see app/analysis/mock.py); the contract is final.",
        "operationId": "analyze_v1_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "analyze"
        ]
      }
    },
    "/v1/health": {
      "get": {
        "description": "Liveness probe.",
        "operationId": "health_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "system"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/events": {
      "get": {
        "description": "Server-Sent Events: the job's state on connect, then on every change, until it ends.",
        "operationId": "job_events_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/videos": {
      "get": {
        "operationId": "list_videos_v1_videos_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VideoOut"
                  },
                  "title": "Response List Videos V1 Videos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Videos",
        "tags": [
          "videos"
        ]
      },
      "post": {
        "description": "Store an uploaded video and register it.",
        "operationId": "upload_video_v1_videos_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_video_v1_videos_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/v1/videos/{video_id}": {
      "delete": {
        "operationId": "delete_video_v1_videos__video_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Video",
        "tags": [
          "videos"
        ]
      },
      "get": {
        "operationId": "get_video_v1_videos__video_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/v1/videos/{video_id}/file": {
      "get": {
        "description": "Serve the video; supports HTTP Range so the player can seek.",
        "operationId": "stream_video_v1_videos__video_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/v1/videos/{video_id}/jobs": {
      "get": {
        "operationId": "list_video_jobs_v1_videos__video_id__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "title": "Response List Video Jobs V1 Videos  Video Id  Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Video Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Start a job for a video. If one of that kind is already running, it is returned (200).",
        "operationId": "start_job_v1_videos__video_id__jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/videos/{video_id}/thumbnail": {
      "get": {
        "operationId": "video_thumbnail_v1_videos__video_id__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Video Thumbnail",
        "tags": [
          "videos"
        ]
      }
    }
  }
}